Fix: improve scannerctl for using redis as storage.
Check the current feed version and only loads it up if it is empty or outdated.
Currently, it takes too long for running a single script, if scannerctl is called with the paramter --path.

Also, add the command line  option to pass the redis url.

+++ rust/redis-storage/src/connector.rs
@@ -707,9 +707,12 @@ impl CacheDispatcher<RedisCtx> {
     pub fn as_dispatcher(
         redis_url: &str,
         selector: &[NameSpaceSelector],
+        force_flush: bool,
     ) -> RedisStorageResult<PerItemDispatcher<CacheDispatcher<RedisCtx>>> {
         let cache = Self::init(redis_url, selector)?;
-        cache.flushdb()?;
+        if force_flush {
+            cache.flushdb()?
+        };
         Ok(PerItemDispatcher::new(cache))
     }
